+++ containers/firecrawl-playwright-cloakbrowser/cloakbrowser_helper.js
@@ -1,7 +1,0 @@
-async function launchCloak(options) {
-  const load = new Function("specifier", "return import(specifier);");
-  const mod = await load("cloakbrowser");
-  return mod.launch(options);
-}
-
-module.exports = { launchCloak };